+++ subsys/net/ip/tcp.h
@@ -535,6 +535,59 @@ int net_tcp_listen(struct net_context *context);
  */
 int net_tcp_update_recv_wnd(struct net_context *context, s32_t delta);
 
+/**
+ * @brief Initialize TCP parts of a context
+ *
+ * @param context Network context
+ *
+ * @return 0 if successful, < 0 on error
+ */
+int net_tcp_get(struct net_context *context);
+
+/**
+ * @brief Unref TCP parts of a context
+ *
+ * @param context Network context
+ *
+ * @return 0 if successful, < 0 on error
+ */
+int net_tcp_unref(struct net_context *context);
+
+/**
+ * @brief Accept TCP connection
+ *
+ * @param context Network context
+ * @param cb Accept callback
+ * @param user_data Accept callback user data
+ *
+ * @return 0 on success, < 0 on error
+ */
+int net_tcp_accept(struct net_context *context, net_tcp_accept_cb_t cb,
+		   void *user_data);
+
+/**
+ * @brief Connect TCP connection
+ *
+ * @param context Network context
+ * @param addr Remote address
+ * @param laddr Local address
+ * @param rport Remote port
+ * @param lport Local port
+ * @param timeout Connect timeout
+ * @param cb Connect callback
+ * @param user_data Connect callback user data
+ *
+ * @return 0 on success, < 0 on error
+ */
+int net_tcp_connect(struct net_context *context,
+		    const struct sockaddr *addr,
+		    struct sockaddr *laddr,
+		    u16_t rport,
+		    u16_t lport,
+		    s32_t timeout,
+		    net_context_connect_cb_t cb,
+		    void *user_data);
+
 #else
 static inline struct net_tcp *net_tcp_alloc(struct net_context *context)
 {
@@ -717,6 +770,47 @@ static inline int net_tcp_update_recv_wnd(struct net_context *context,
 {
 	ARG_UNUSED(context);
 	ARG_UNUSED(delta);
+
+	return -EPROTONOSUPPORT;
+}
+
+static inline int net_tcp_get(struct net_context *context)
+{
+	ARG_UNUSED(context);
+
+	return -EPROTONOSUPPORT;
+}
+
+static inline int net_tcp_unref(struct net_context *context)
+{
+	ARG_UNUSED(context);
+
+	return -EPROTONOSUPPORT;
+}
+
+static inline int net_tcp_accept(struct net_context *context,
+				 net_tcp_accept_cb_t cb, void *user_data)
+{
+	ARG_UNUSED(context);
+	ARG_UNUSED(cb);
+	ARG_UNUSED(user_data);
+
+	return -EPROTONOSUPPORT;
+}
+
+static inline int net_tcp_connect(struct net_context *context,
+				  const struct sockaddr *addr,
+				  struct sockaddr *laddr,
+				  u16_t rport, u16_t lport, s32_t timeout,
+				  net_context_connect_cb_t cb, void *user_data)
+{
+	ARG_UNUSED(context);
+	ARG_UNUSED(addr);
+	ARG_UNUSED(laddr);
+	ARG_UNUSED(rport);
+	ARG_UNUSED(lport);
+	ARG_UNUSED(cb);
+	ARG_UNUSED(user_data);
 
 	return -EPROTONOSUPPORT;
 }
